update dialog for deleting labels to shadcn

+++ src/routes/(app)/labels/DeleteLabelDialog.svelte
@@ -1,4 +1,7 @@
 <script lang="ts">
+    import LabelDot from '$lib/components/label/LabelDot.svelte';
+    import { Button } from '$lib/components/ui/button';
+    import EncryptedText from '$lib/components/ui/EncryptedText.svelte';
     import { ANIMATION_DURATION } from '$lib/constants';
     import { dispatch } from '$lib/dataChangeEvents';
     import LabelSelect from '$lib/components/label/LabelSelect.svelte';
@@ -10,7 +13,8 @@
     export let id: string;
     export let color: string;
     export let name: string;
-    export let reloadOnDelete = true;
+    export let onDelete: undefined | (() => void);
+    export let cancel: undefined | (() => void);
     export let labels: Record<string, Label>;
 
     let changeLabelId: string;
@@ -21,31 +25,29 @@
                 strategy: 'remove'
             })
         );
-        if (reloadOnDelete) {
-            location.reload();
+        await dispatch.delete('label', id);
+        if (onDelete) {
+            onDelete();
             return;
         }
-        await dispatch.delete('label', id);
-        popup.set(null);
     }
 
     async function reassign() {
+        if (!changeLabelId) {
+            notify.error('Please select a label to reassign to.');
+            return;
+        }
         notify.onErr(
             await api.delete(apiPath(`/labels/?`, id), {
                 strategy: 'reassign',
                 newLabelId: changeLabelId
             })
         );
-        if (reloadOnDelete) {
-            location.reload();
+        await dispatch.delete('label', id);
+        if (onDelete) {
+            onDelete();
             return;
         }
-        await dispatch.delete('label', id);
-        popup.set(null);
-    }
-
-    function cancel() {
-        popup.set(null);
     }
 
     function filter(label: Label) {
@@ -55,74 +57,33 @@
 
 <div>
     <div class="pb-4 flex items-center gap-2">
-        <span class="rounded-full aspect-square w-5 h-5 flex" style="background-color: {color}" />
-        <p class="text-xl"> Delete Label '{name}' </p>
+        <p class="text-xl inline-flex items-center gap-2">
+            Delete Label
+
+            <LabelDot {color} big />
+            <EncryptedText text={name} />
+        </p>
     </div>
     <p> There are entries, edits or events with this label. </p>
-
-    <div class="options" in:slide={{ duration: ANIMATION_DURATION }}>
-        <div>
+    <div class="py-2" in:slide={{ duration: ANIMATION_DURATION }}>
+        <div
+            class="flex items-center justify-between rounded-xl p-2 border border-border border-solid"
+        >
             <LabelSelect bind:value={changeLabelId} {filter} {labels} />
-            <button on:click={reassign}>
-                Give Different Label to Entries/Events with this Label
-            </button>
+            <Button on:click={reassign}>Give different label</Button>
         </div>
 
-        <h2>OR</h2>
+        <h2 class="py-4">OR</h2>
 
-        <div>
-            <button on:click={rmLabel}> Delete Label and Remove from Entries/Events </button>
+        <div
+            class="flex items-center justify-between rounded-xl p-2 border border-border border-solid"
+        >
+            <p />
+            <Button on:click={rmLabel}>Remove label</Button>
         </div>
 
-        <div class="cancel">
-            <button on:click|self={cancel}> Cancel </button>
+        <div class="pt-8">
+            <Button variant="outline" on:click={cancel}>Cancel</Button>
         </div>
     </div>
 </div>
-
-<style lang="scss">
-    .options {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        margin-top: 2rem;
-
-        & > div {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            border: 1px solid var(--border-color);
-            border-radius: $border-radius;
-            padding: 0.6em;
-            margin: 0.5em;
-
-            button {
-                border-radius: $border-radius;
-                margin: 0.5rem;
-                padding: 0.3rem;
-
-                &:hover {
-                    background: var(--accent-danger);
-                }
-            }
-
-            &.cancel {
-                border: none;
-
-                button {
-                    padding: 0.5rem;
-
-                    &:hover {
-                        background: var(--light-accent);
-                    }
-                }
-            }
-        }
-    }
-
-    button {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-</style>
